Imported Upstream version 3.2.2-dfsg

+++ src/VBox/Devices/VirtIO/Virtio.cpp
@@ -1,4 +1,4 @@
-/* $Id: Virtio.cpp 28907 2010-04-29 16:09:43Z vboxsync $ */
+/* $Id: Virtio.cpp 29917 2010-05-31 16:05:25Z vboxsync $ */
 /** @file
  * Virtio - Virtio Common Functions (VRing, VQueue, Virtio PCI)
  */
@@ -50,7 +50,7 @@ static void vqueueReset(PVQUEUE pQueue)
 
 static void vqueueInit(PVQUEUE pQueue, uint32_t uPageNumber)
 {
-    pQueue->VRing.addrDescriptors = uPageNumber << PAGE_SHIFT;
+    pQueue->VRing.addrDescriptors = (uint64_t)uPageNumber << PAGE_SHIFT;
     pQueue->VRing.addrAvail       = pQueue->VRing.addrDescriptors
         + sizeof(VRINGDESC) * pQueue->VRing.uSize;
     pQueue->VRing.addrUsed        = RT_ALIGN(
